+++ apps-client/financas/free/changelog.md
@@ -6,6 +6,7 @@ Changelog
 Features
 - Remover campo número da nota
 - Inserir campo observações
+- Remover número de parcelas de transferencia
 
 ## 2.1 (2013-04-15)